requirements.txt
discord.py>=2.4.0
aiosqlite>=0.20.0
PyYAML>=6.0.2

